finagle-toggle: Hash the inputs to ToggleMap Toggles
Problem

It is too easy for users to have inputs with poor distributions
into the `Toggles` created by a `ToggleMap` which can lead to
unexpectedly bad distributions of the fractional toggles.

Solution

Apply a murmur3 hash to the input to give a better distribution.

Result

This allows even poorly distributed values to be used as inputs
to these `Toggles`.

RB_ID=899195