util-core: Standardize argument checking in Buf.write,slice
Problem

The implementations of `c.t.io.Buf.write` and `Buf.slice` had a bunch
of duplication along with a bunch of slight differences. Additionally,
some of implementations were using the 2-argument version of
`Predef.require` which allocates a `Function0`. And, implementations
did not always include the input argument that was invalid in the
failure message.

Solution

Factor out the logic into protected methods on `Buf` and then use
them.

Result

More consistent behavior.

RB_ID=899935