finagle-core: Back Context with a Map and reduce the API surface
Problem

The structure of c.t.f.context.Context is a linked list of
cons cells and tombstones that is never reduced, leading to
the possibility of unbounded collection growth. This can result
in memory leaks and StackOverflowErrors since the traversal
functions are not tail recursive.

Solution

Change the internal representation to an immutable Map which has
effectively constant update and access operations. Benchmarks show
this to be slightly slower than the linked list structure (~15%, 10 ns)
and creates a bit more garbage than for inserts/shadows (~20 bytes/op)
and slightly less for clears, but the reduction in complexity is
well worth it since these are relatively rare operations.

RB_ID=896663