finagle-core: InetResolver can now be created with a custom FuturePool
Problem

The InetResolver by default uses a unbounded threadpool under the hood.
This can be a problem for applications that want to control over number
of threads.

Solution

Added optional FuturePool argument to the InetResolver factory.
Applications can now create a custom Resolver which uses delegation and
a custom InetResolver.

Signed-off-by: Bryce Anderson <banderson@twitter.com>

RB_ID=901969