finagle-http: Remove Message.write(ChannelBuffer) method
Problem

The `Message.write(ChannelBuffer)` method is netty3 specific and
must go in order to present a netty free API. In addition, the
semantics of the different write methods are very different and
a may have incorrect comment strings (specifically withOutputStream).

A further problem is the methods behavior when the message is chunked:
the base `.write(ChannelBuffer)` method, used by many other methods,
checks to see if the message is chunked, and if it is, it attempts to
write it to the messages `Writer`. This poses a few problem:

* If the `Writer` is busy this throws an exception. There is no way to
  probe the state of the `Writer` other than through catching
  exceptions.
* There is no reasonable way to close the `Writer` since we don't have
  a handle to the `Future` generated by the write which signals when
  the previous operation is complete.

Solution

Remove the `Message.write(ChannelBuffer)` and replace it with an
overload `Message.write(Buf)` with the same semantics for non-chunked
messages and throws an `IllegalStateException` if the message is
chunked.

RB_ID=900091