finagle-http2: Allow initial settings to Http2Codec
Problem:
While possible when creating a Finagle client, there is currently
no way to specify initial Http/2 settings for the Http2Codec when
creating a server in Finagle.

Solution:
Netty 4.1.7 introduces a way to specify settings when instantiating
an Http2Codec. Thread the existing Settings Param through to the
places where this happens.

Result:
It is now possble to specify initial Http/2 settings for Finagle
servers

RB_ID=903667