finagle-http: Introduce HeaderMap.getOrNull
Problem

Preserving Netty behavior w.r.t. retrieving header values is hard in Java since
there is no easy way of getting a header or `null` if it does not exist. In Scala,
this could be mimicked with `get("foo").orNull`, which is not accessible from Java
(think of implicit param for `orNull`). It's still possible to `getOrElse` from
Java but it requires pre-allocating a `Function0` instance that returns a default
value (this is painful).

Solution

Introduce Java-friendly `headerMap.getOrNull` method.

RB_ID=904093