finagle-http: Don't set Transfer-Encoding to chunked when Content-Length present
Problem

There is a discrepency between Netty 3 and Netty 4 Finagle
client behavior. If a request has content-length set in N3,
and is chunked (but does not have the Transfer-Encoding header
set), it will *not* be sent with Transfer-Encoding: chunked.

In Netty 4, however, if a request is chunked, the request is sent
with Transfer-Encoding: chunked and Content-Length is set to 0.
This causes an issue when servers examine the content body of the request
and find it empty.

Solution

Only set the Transfer-Encoding to "chunked" when a request does not
have a content length.

RB_ID=903657