util-core: Introduce Buf.Indexed for positional access to Bufs
Problem

The `c.t.io.Buf` abstraction lacks a cheap mechanism to do indexed
access to the underlying bytes. Without this, the code relies on
access to a byte-array being a cheap operation. While this is often
the case today, it is not always the case, and is an impediment to
moving towards off-heap buffers.

Solution

Introduce the `Buf.Indexed` abstraction that adds a positional API
`apply` and a scanning API `process`. Existing `Buf` implementations
in util and finagle are updated to implement the API.

Result

A benchmark comparing `slice` and `write` to `apply` shows the
performance differences, and this is done giving slice and write an
advantage by reusing the byte-array. Future commits will put the API
to use.

RB_ID=904559