finagle-netty4: Netty4ProxyConnectHandler should not buffer writes
Problem

While `NettyProxyConnectHandler` buffers writes it doesn't properly drain them
when the proxy handshake is finished. Instead of removing the HTTP pieces from the
pipeline and then draining writes it does it in the reverse order, which leads
to TLS/SSL (when configured) handshake hello message beeing dropped on the floor
(by HTTP codec).

Solution

Do not buffer writes on the `NettyProxyConnectHandler` and let the `ProxyHandler`
(placed right after) do this job.

RB_ID=904803