finagle-http: Add content(Buf) to Message
Problem

It's difficult to use our http types from Java because many of our
mutator methods are of the `name_=(..)` style. This is true for the
`Message` content as well. In addition, our only existing
`setContent(ChannelBuffer)` method is package protected but easy to use
from Java so it's being heavily used outside of finagle, hindering the
migration to Netty 4.

Solution

Deprecate the `getContent: ChannelBuffer` and `setContent(ChannelBuffer)`
methods and add a new method, `content(Buf)` which should be easier to
use from Java but uses the appropriate twitter types.

RB_ID=904749