finagle-http: Deprecate http Request Response protected[finagle] methods
Problem / Solution

Most of these methods deal in Netty 3 specific types which is hindering
the transition to Netty 4, so they should be deprecated then removed.
The `setUri(String)` method of `Request` is also deprecated and a public
`uri(String)` method is added to offer a reasonable public API for Java
users.

RB_ID=905761