finagle-http: Add client exception TooLongMessageException
Problem

We don't always handle important netty exceptions in finagle,
such as the TooLongFrameException on the client side, so folks
need to do that themselves. Unfortunately this is a leak of the
underlying Netty implementation, and that makes it hard to move
seamlessly from Netty 3 to Netty 4, since they both offer an
exception of the same name, but they are of different type because
they are from totally different packages.

Solution

Catch these Netty exceptions in the Netty pipeline and wrap them
in a named finagle ChannelException, so the backend no longer
matters.

As a bonus, I fixed a bug in the netty4 client where fixed length
messages were bounded to the `maxChunkSize` instead of the appropriate
`maxResponseSize`.

RB_ID=905567