finagle-http2: Re-enable failfast
Problem

We use exceptions for flow control in the MultiplexedTransporter when we close
the connection during normal operation.  We had to remove FailFast so that it
wouldn't be triggered by this, so http/2 doesn't use FailFast right now.

Solution

Return a dead service when the connection is dead, and make sure that clients
know to interpret that as a cue to retry.

Result

We can turn on fail fast again in http/2!

RB_ID=905445