Problem
We do a number of unchecked casts of our Transport[In, Out] type, which is
due to Java type erasure, that can lead to `ClassCastException`s far away
from where the cast actually happened, making it difficult to debug.

Solution

Require a Class[T] when performing the cast, which we can use to verify that
the instance in question is assignable as a T. This results in fail-fast
behavior.

RB_ID=902053