finagle-http: Deprecate protected Message.headers method
Problem

We are exposing the netty3 `HttpHeaders` type, which is
making it impossible to fully upgrade to netty4. In addition,
its scoped protected, which means Java users see it as public
and then it escapes our control.

Solution

Deprecate it and remove all usages. In the process, the `getAll(key)`
method has been changed to return a `Seq[String]` as opposed to an
`Iterable[String]`.

RB_ID=905019