finagle-base-http: Refine c.t.f.http.Method
Problem

Our model of the HTTP method uses sub-objects of the abstract Method class
for common method names, but they have no different behavior, meaning
its just a bunch of useless .class files and artificial subtype
relationships. Additionally, the only way to get the name of the HTTP
method is from the `.toString` method, which is generally reserved for
debugging and, in general, shouldn't be used for business logic.

Solution

Make a single final Method class that has a `.name` method. Common HTTP
methods are just cached instances of the same Method class, and getting
a Method instance is only possible by using a cached instances or the
`Method.apply(String)` factory function.

RB_ID=906697