finagle-base-http/finage-netty4-http: Fix HeaderMap descendants
Problem

The behavior of the descendants of HeaderMap are not self-consistent.
More specifically, the Netty4HeaderMap was inconsistent with its
key iterators/set and the behavior of `+=`, which was a synonym for
`.add(..)` as opposed to `.set(..)` for the other HeaderMap impls.
MapHeaderMap behavior of `+=` also had some weird quirks that set
it apart from the Netty3HeaderMap.

Solution

Consolidate the HeaderMap tests into one abstract unit,
AbstractHeaderMapTest, which can be used to check the behavior
of all the HeaderMap implementations. Next, fix the things that
are broken.

RB_ID=906406