finagle-native: Remove Certificate and Key Generation from Tests
Problem / Solution

Finagle-Native uses openssl to generate certificates from the command
line when running tests. This is unnecessary, as predefined certs and
keys as resources will work equally well. Also, the format which is
generated for the private key is not one which will be supported going
forward, as it is not PEM-encoded PKCS#8. This changes the tests to use
predefined certificates and keys.

RB_ID=907115