finagle-netty4-http: Respect decoder result in HTTP clients
Problem / Solution

Not all Netty 4 HTTP pieces fire an exception into the pipeline when
something bad happens. Some just set the `decoderResult` on the inbound
message to a failure. We already respect that on a server-side and have
appropriate checks/converters in place for malformed HTTP requests. We
should do the same things for clients.

RB_ID=907096