util-core: Promote Buf.Indexed API
Problem

The `c.t.io.Buf.Indexed` API is marked as `private[twitter]` but it is
generally useful for end-users.

Solution

Promote it to be part of `Buf`. This required renaming the indexed
byte lookup method to `get` from `apply` to avoid the naming conflict
with the `Buf.apply` on the companion object.

RB_ID=907231