finagle-netty4: Stop setting SO_LINGER=0
Problem

Not all Channels support setting SO_LINGER and as such Netty 4 logs a
warning when the option is set on channels without support.

Solution

While we could push this setting down to after the `Channel` is
established, we could not find a good reason to keep this setting,
which appears to be simply carried over from our Netty 3
implementation.

Result

SO_LINGER is no longer set to 0 for Netty 4 servers.

RB_ID=907325