finagle-http2: Adds support for nacking
Problem

finagle-http uses headers to signal that it wants to nack, but
http/2 can do better by using RSTs and normal http error codes.

Solution

We add a handler which converts from the special headers into RSTs.
Then, on the client side, we ensure that those RSTs are handled
properly.  As a side effect of this change, it means that if a
server nacks a client that has a connection pool of size one, which
is the common case, its connection pool will drop to size zero and
the connection will be closed, since connections are ref-counted.

Instead of ref-counting streams, we instead say that connections
won't be closed until the client asks for them to close, and make
Transporters Closable.  This seems a little strange because
Transporters ask for a SocketAddress, so it seems like closing a
Transporter would close all connections.  However, this is a legacy
of the old model of Transporters–Transporters are now created newly
for each socket address, so no two socket addresses share the same
Transporter.  This is important for LatencyCompensation to adjust
TCP Connect Timeouts correctly.  We will clean up the Transporter
API in the future.

In the mean time, we provided a default implementation for close
and ensure that Http2Transporter specially closes the underlying
transport if one exists, and it has upgraded to http/2.  Importantly,
it doesn't matter in the downgrade case, because those transports
can be closed directly instead.

Result

Finagle's http/2 implementation is now properly hooked up with its
NACKing API.

RB_ID=907292