util/util-tunable: Introduce TunableMap for accessing Tunables
Problem

Because users will not be creating Tunables directly to use as
configuration parameters for Finale clients/servers, we need
a mechanism for accessing Tunables.

Solution

Create a TunableMap that maps keys containing Tunable ids and types to Tunables.
Tunables retrieved from this map see updates to their values when the values
of the Tunables within the map are updated.

RB_ID=906547