finagle-docs: Add User & Developer guide for Exceptions and Failures
Problem

Lots of the information about Exceptions and Failures is scattered in folks
memories and notebooks, making it difficult for both users and developers
to know what to do.

Solution

Add official documentation. Also, I've moved Extending in to the
developer/ docs and added developer/Futures.rst to the TOC.

RB_ID=904533