finagle-core/finagle-mux: Force sized ByteArrays for context entries
Problem

Our MarshalledContext attempts to copy ByteBufferBuf types into the heap
in an effort to avoid keeping a strong reference to a large buffer only to
expose a small view of them. This is Netty 3 specific and won't be helpful
for Netty 4 users.

Solution

Force the Buf to a ByteArray with the exact sized backing Array[Byte],
making a copy if necessary, which is expected to be the common case for Mux.
This logic is moved to the Mux decoder where we have the scope to know that
the Buf represents a view of the larger message. Other protocols that use
MarshalledContext (vanilla thrift, HTTP) are already doing the right thing
and therefore doesn't need modification.

RB_ID=907663