finagle-http2: Notice when remote peers cut connections
Problem

Our cleartext clients disable autoread when they fail the http/2 client upgrade,
which is a state transition that wasn't accounted for in ChannelTransport.  This
means that ChannelTransport's bookkeeping is incorrect.

Solution

Always do the bookkeeping in ChannelTransport, but only trigger the extra read
when autoread is enabled.  Also, cut the connection when a read encounters an
error in MultiplexedTransporter.

Result

Http/2 clients eagerly notice when the remote peer has disconnected, even when
they downgrade to http/1.1.

RB_ID=908081