finagle-core: Make Transporter coherent
Problem

We construct a new Transporter for each remote host to support different latency
compensation for each name.  However, Transporter takes a socket address when
constructing a new Transport, which makes it seem like we construct a single
Transporter per client.

Solution

In order to make this more clear, make Transporter no longer take SocketAddress
as an argument.  Instead, StackClient directly injects it into the new Transporter.

In general the edits were small, but in finagle-http2, which was written with
the assumption that Transporters were shared across connections, we're able to
change from using a ConcurrentHashMap to an AtomicReference, which represents a
substantial simplification of the model.

Result

The Transporter api now reflects the way it actually works.

RB_ID=907544