finagle-http: Do not validate headers twice when N4 is enabled
Problem

When Netty 4 is enabled in finagle-http, headers are validated twice: first by
N3 `DefaultHttpHeaders` (when a Finagle message is being constructed) and then
by N4 `DefaultHttpHeaders` (when a Finagle message is being converted to N4).
This is not only inefficient but also introduces a behavioral difference between
N4 and N3 implementations: headers with null-chars were allowed in N3 but prohibited
in N4.

Solution

To preserve the N3 behavior and not perform the same work twice, we're skipping
headers validation in the N4 bijections since they were already validated by N3
at that point.

RB_ID=908368