finagle-http: Clients don't force close the connection on 1xx, 204, or 304
Problem

If a finagle-http client receives a 1xx, 204, or 304 response,
which typically don't include transfer-encoding or content-length
headers because they are defined to not have a body in RFC 7230,
the client forces the connection closed because it believe the
response body must be delimited by the end of connection.

Solution

Check to see if the status code is one that must not have a body
when determining if the connection must be closed.

RB_ID=908593