finagle-core: SSL/TLS Param Changes for Finagle on Netty 3
Problem

SSL/TLS params, configurations, and code are out of date, not
customizable, and inconsistent between versions.

Solution

Replace the SSL/TLS params for Finagle on Netty 3 with separate
configuration and engine creation parameters which are extensible. These
params will soon be migrated to Netty 4 as well for consistency.

RB_ID=907923