finagle-http2: Added support for ALPN
Problem

In practice, http/2 is only implemented on the web by clients and servers that
speak over TLS.  They use ALPN, which isn't easily accessible in finagle.

Solution

Add a finagle-internal way of configuring ALPN on a per-protocol basis, and hook
up the finagle-http2 upgrade path to ALPN on the client-side.

Result

finagle-http2 clients and servers that are set up for TLS will handle ALPN
properly now!

RB_ID=908310