finagle-http: Initialize connection to be KeepAlive
In ConnectionManager, the connection starts off with "isKeepAlive = false".
Because the connection is initially idle, ConnectionManager returns "true"
for shouldClose(), which is called by Retries before a request is made (if the
client is constructed without FactoryToService) to get the status of the connection.
We don't want a connection to start off in a closed state because service acquisition
will be retried until exhausted.

Solution

Initialize "isKeepAlive" to "true".

RB_ID=908622