ZooKeeper exists watch now ignores session state events
Problem

When a ZooKeeper "exists" watch is triggered because the path being
watched is created or deleted, ZooKeeper removes that watch, and it has
to be re-added in order to receive further notifications about the
watched path.

However the exists watch also gets triggered for ZooKeeper session events
(connected, disconnected etc) - for those events, ZooKeeper does not remove
the watch, so it doesn't have to be re-added.

Currently the code that re-adds the watch does not distinguish between
event types and always re-adds a new watch, causing an ApacheWatcher
memory leak each time the watch is triggered for a session event type.

Solution

Filter out session event types when watching for path existence