scrooge, finagle-thrift, finatra: Revisit ServiceIface API
Problem

The ServiceIfaces generated by Scrooge for Scala's Service
per-endpoint are a c.t.f.Service[ThriftMethod.Args, ThriftMethod.Result].
Unfortunately, Result here is is not a very pleasant interface to
work with as they are a ThriftResponse[ThriftMethod.Success] and
this means you have to do some odd pattern matching to use them.

For example, using an example IDL:

exception AnException {
  1: i32 errorCode
}
exception AnotherException {
  1: i32 errorCode
}
service SomeService {
  i32 TheMethod(
    1: i32 input
  ) throws (
    1: AnException ex1,
    2: AnotherException ex2
  )
}

This is roughly what pattern matching against a
SomeService.TheMethod.Result looks like:

result match {
  case Return(SomeService.TheMethod.Result(Some(v), _, _))) =>
    // Successful
  case Return(SomeService.TheMethod.Result(_, Some(e), _))) =>
    // AnException
  case Return(SomeService.TheMethod.Result(_, _, Some(e)))) =>
    // AnotherException
  case Throw(_) =>
    // an exception not from Thrift
}

Solution

The ServiceIface is better modelled as a Service from ThriftMethod.Args
to ThriftMethod.SuccessType. This is far more idiomatic as the result
of the Service is a Future[SuccessType] and thus it already is a Try.

This turns that pattern match into roughly:

result match {
  case Return(success) =>
  case Throw(AnException(_)) =>
  case Throw(AnotherException(_)) =>
  case Throw(_) =>
}

Result

While this is a breaking API change, it should be fairly superficial.
Translating from a Result to a SuccessType is fairly
straightforward and an example can be seen in Scrooge's
ThriftServiceIface.resultFilter.

RB_ID=909434
TBR=true
NO-QUEUE=true