finagle-base-http: Remove most protected[finagle] methods from HTTP Response
Problem

The Finagle HTTP model includes a number of protected[finagle]
methods exposing the underlying Netty 3 types that back the model,
keeping us coupled to Netty 3. These can also be used from Java
and make it difficult encapsulate the Netty 3 implementation
details.

Solution

Remove most of these or make them legitimate protected so that
they can easily be removed later as we transition away from a
HTTP model which is a proxy to a Netty type.

RB_ID=908409
TBR=true