finagle-core: divorce Aperture's load band from a node's load metric
Problem

The `LoadBand` trait implemented a Node in order to keep track
of the total number of pending requests. This made it less reusable and,
in fact, it need not know anything about a Node type.

Solution

Create a new entry point in `Balancer` that allows traits to have access
to the underlying `ServiceFactory`. This allows `LoadBand` to do the
neccessary bookkeeping to track the total number of pending requests.

RB_ID=909474