finagle-core: Refactor MethodBuilder
Problem

In order to have a useful `MethodBuilder` for Thrift/ThriftMux, which
have different request and response types per-endpoint, the APIs
needed some refactoring.

Solution

`MethodBuilder` now works in terms of `Filter.TypeAgnostic` so
that the filters will be usable by Thrift and ThriftMux clients.

Result

This clears the path for a future commit that will allow Thrift and
ThriftMux clients to utilize `MethodBuilder`.

RB_ID=909497