finagle-http: Server errors now considered failures
Problem

The HTTP protocols (1.1 and 2) treat responses with 500 status codes
as successful for the purposes of metrics and failure accrual. This is
surprising to users.

Solution

Change the behavior to classify these as unsuccessful. Users can opt
out of this via the
`com.twitter.finagle.http.serverErrorsAsFailuresV2` toggle.

RB_ID=909315