finagle-http: Introduce MethodBuilder for HTTP
Problem

HTTP clients should be able to use the MethodBuilder API.

Solution

Introduce `Http.Client.methodBuilder` as the entry point.

Result

Users get an API with retries, timeouts, and logical success metrics.
As an example, this produces a `Service` that has timeouts and retries
on a 418 status code.

    import com.twitter.conversions.time._
    import com.twitter.finagle.Http
    import com.twitter.finagle.service.{ReqRep, ResponseClass}
    import com.twitter.util.Return

    val client: Http.Client = ???
    client.methodBuilder("inet!example.com:80")
     .withTimeoutPerRequest(50.milliseconds)
     .withTimeoutTotal(100.milliseconds)
     .withRetryForClassifier {
       case ReqRep(_, Return(rep)) if rep.statusCode == 418 => ResponseClass.RetryableFailure
     }
     .newService("an_endpoint_name")

RB_ID=909793