util-core: Add method to Buf for efficient writing to nio ByteBuffer
Problem / Solution

We often use Buf in the context of doing IO operations, and the Java
asynchronous API's deal in nio ByteBuffers, and in particular, the IO
operations will require copying the data to a direct nio ByteBuffer. To
make this simpler, we add a method to Buf to write its contents to a nio
ByteBuffer.

RB_ID=910152