finagle-core: Add ConnectionRegistry to track which clients are connected to a finagle server
Problem

There is currently no way to see which clients are connected to a finagle
server. This would be useful information for monitoring clients that create
too many connections and other similar issues.

Solution

Add a registry that will track the remote address of a client and when the
connection was established.

RB_ID=909656