finagle-core: Fix ConnectionRegistry tests
Problem

The ConnectionRegistry tests in StringServerTest don't account for how the
ConnectionRegistry is global and can contain entries before the test starts.
This can cause the tests to fail.

Solution

Capture the initial state of the registry to ensure the ConnectionRegistry test
isn't affected by other tests that may also register.

RB_ID=910375