finagle-netty4: Use channel's allocator on the outbound path of BufCodec
Problem

We're almost ready to enable pooling and want to make sure we'll take an advantage
of it when it's finally here.

Solution

Using current channel's allocator in `BufCodec` (outbound path) would allow us to
not only benefit from pooling (when enabled) but also reduce allocations needed to
send `Buf.Composite` messages (literally, every Finagle message).

There is no zero-copy converters available from `Buf.Composite` to `CompositeByteBuf`
and it's not that trivial to write those. However, it's possible to reduce the allocations
on the outbound path for `Buf.Composite` knowing a bit of details on how Netty handles
the outbound buffers.

In other words, this change serves two purposes:

 - Prepares the ground for pooled allocators
 - Improves the performance of Finagle clients/servers sending composite `Buf`s
   over the wire (ThriftMux does that)

RB_ID=910000