Problem
A GOAWAY currently clears all of the HTTP2 streams queues instead of following
the spec.

Solution

Examine the last stream ID in the GOAWAY message and close any streams with
higher IDs. Allow streams with lower or equal IDs to finish operation. This
allows the GOAWAY message to function as intended and gracefully shut down
client connections. Once a GOAWAY message is received, the parent
MultiplexedTransporter is marked as dead and no new streams can be created.

RB_ID=910086