finagle-core: Add windows IOException string to FinestIOExceptionMessages
Problem

When running the service on Windows and the client socket is closed from
the client side the log is being polluted with WARNINGS containing
exceptions: "An existing connection was forcibly closed by the remote
host".

Solution

This exception message should be added to the existing list of messages
to be logged with the FINEST level.

Result

adding windows specific IOException that indicates that the socket was
closed externally to be logged with FINEST level.

Signed-off-by: Bryce Anderson <banderson@twitter.com>

RB_ID=910760