finagle-http2: Fix PriorKnowledgeHandler ByteBuf slicing
Problem

The PrioKnowledgeHandler is using `buf.capacity() - buf.readerIndex()`
as the length to slice when it should be using `buf.readableBytes()`.

Solution

Make it the next handlers problem and don't slice the ByteBuf and just
pass the ByteBuf along as is. This also fixes a potential memory leak
where an empty ByteBuf was not released.

RB_ID=911519