finagle-core: Fix MethodBuilder's default timeouts
Problem

The docs for `MethodBuilder` timeouts were a lie and claimed that
there were no defaults. However, in practice users would get the
configuration from the underlying `StackClient`. While working on
adding tests for this, I uncovered a bug with regards to disabling
timeouts on a `MethodBuilder`.

Solution

The docs now reflect the reality and users can now disable timeouts.

RB_ID=911761