finagle-core: Support mutable Stack.Param values in StackRegistry
Problem

StackRegistry.Entry.modules is an immutable data structure that contains
the names and values of Stack.Params. For mutable Stack.Params, this
means that the param values in StackRegistry.Entry.modules do not reflect
changes.

Solution

Make StackRegistry.Entry.modules a function that gets a snapshot of the
current stack and param values.

RB_ID=911713