finagle-core: Simplify client construction with endpoint clients
Problem

Our hands are somewhat tied with how to implement an endpoint, because we assume
that all implementations use Netty and that Netty is consistent, neither of
which is always the case.

Solution

Provide a helper class, EndpointerStackClient, that doesn't need to use a
Transporter and Dispatcher.  Among other things, this lets us simplify our
implementation of h2, so we can make Transporter not extend Closable anymore!

Result

It's easier to add custom logic at the endpointer layer.

RB_ID=912889