finagle-http2: Liveness detection via pings
Problem

We want support for the resiliency feature from mux in http/2 too.

Solution

Hook up the FailureDetector to pings on the client-side.  Netty responds to
pings automatically, so we don't need to do it explicitly on the server-side.

Result

finagle's http/2 implementation can take advantage of PINGs

RB_ID=913341