finagle-core/finagle-netty4: Transporter implementations emit better error on connect failure
Problem

We have a minor discrepancy in the Netty 3 and Netty 4 Transporter
implementations: on failure to establish a connection, the Netty 4
implementation returns a Failure.rejected(String) with the
underlying exception wrapped while Netty 3 returns a WriteException.

Solution

Change both implementations to return a
Failure(ConnectionFailedException). In addition, The exception
emitted when connection establishment is canceled was changed from
a WriteException to a CancelledConnectionException wrapped in a
Failure.

RB_ID=912113