finagle-http2: Reduce log level on interrupt
Problem

We now log loudly if a connection fails the http upgrade.  However, sometimes it
fails the upgrade for a normal reason, like it's interrupted while trying to
establish a connection, or the remote peer cuts the connection.  We typically
don't want to log in this case.

Solution

Provide a guidance log level to log at a better level in the common case.

RB_ID=913743