finagle-http2: Strip out headers found in the h1 "connection" header
Problem

Connection-related headers should be stripped from h2 requests masquerading as
h1 requests for compatibility. According to
https://tools.ietf.org/html/rfc7230#section-6.1, the names of other connection
related headers may be specified in the "connection" header. These should also
be stripped, but are not.

Solution

Subclass Http2ServerDowngrader and intercept the header conversion. Remove all
headers that are referenced in "connection" and handle the "TE" header as per
the spec.

RB_ID=913799