finagle-toggle, twitter-server: Capture the last value of Toggle
Problem

For toggles that are fractionally enabled, it is hard for developers
to know the state of their toggle across their cluster.

Solution

Introduce `Toggle.Captured` which is used by the standard toggle
mechanisms. This captures the last value of `Toggle.apply` and can be
introspected.

Result

Users can introspect TwitterServer's /admin/toggles endpoint to see
the lastValue.

For example, here is a server running with
`com.twitter.finagle.netty4.poolReceiveBuffers` enabled and lastValue
appears as true.

    $ curl -s http://0.0.0.0:7090/admin/toggles/com.twitter.finagle.netty4/com.twitter.finagle.netty4.poolReceiveBuffers
    {
      "libraries" : [
        {
          "libraryName" : "com.twitter.finagle.netty4",
          "toggles" : [
            {
              "current" : {
                "id" : "com.twitter.finagle.netty4.poolReceiveBuffers",
                "fraction" : 1.0,
                "lastValue" : true,
                "description" : "Enable pooling of receive buffers on Netty 4"
              },
              ...

RB_ID=913925