finagle-serversets: Fix AddressOrdering
Problem

It's possible for `AddressOrdering` to fail when passed into the
default java util sort methods since it is inconsistent across
comparisons. For example, transitivity doesn't hold with the
following example if we fallback to sorting by octets when
the shardId is missing:

a: shardId: Some(2), octects: 1.2.3.3
b: shardId: None, octects: 1.2.3.4
c: shardId: Some(1), octects: 1.2.3.5

a < b
b < c
a > c (non-transitive ordering!)

In general, it's dangerous to compose orderings like this.

Solution

Remove the `fallback` ordering in `ZkMetadata.AddressOrdering`,
and handle all the cases to compare two addresses with varying
metadata.

RB_ID=914111