finagle-core: simplify the AddressOrdering param and log on failure
Problem / Solution

It's possible for the activity pipeline in between TrafficDistribtor
and building a new distributor in `Balancer` to fail. We want to log
when this happens so it's easier to debug. This patch also simplifies
the interface for AddressOrdering to no longer take a StatsReceiver
since orderings were simplified.

RB_ID=914113