finagle-http2: Stop upgrading before request has finished
Problem

When we send an h2c upgrade request, we can receive an upgrade response
immediately, but we shouldn't upgrade the pipeline until we've finished sending
the rest of the h2c upgrade request.

Solution

When the upgrade is successful, delay upgrading until after the initial upgrade
request has finished.

RB_ID=914104