finagle-core: set min aperture when number of clients < servers
Problem

In cases where there are fewer clients than servers, we want to
additionally ensure that we get full ring coverage. For example,
imagine the degenerate case where we have an min aperture of size
1 and fewer clients than servers – we know that we will at most cover
`numClients` of the server ring. Thus, we need a min aperture size
of at least ceil(`numServers` / `numClients`).

Solution

Calculate the min aperture in this case.

RB_ID=914322