finagle-core: Hoist logic from StdStackServer into an intermediate, ListeningStackServer
Problem

The StdStackServer is based on the transport + disaptcher model for
establishing connections and dispatching requests. In addition, it
contains most of the logic used for tracking sessions and registering
connections. This makes it difficult to build servers that don't rely
on the transport + dispatcher model.

Solution

Create a new intermediate type, ListeningStackServer, which acts as an
implemenentation template that isn't dependant on a specific session
pattern. The logic from StdStackServer responsible for preparing the
ServiceFactory and tracking sessions is hoisted into this new template.

RB_ID=914124