finagle-core: Fix Retries docs and add test for svc acquisition requeues
Problem

The docs for c.t.f.service.Retries suggest that service
application requeues are gated by `RequeueFilter` (and thus
a budget), but this is not the case.

Solution

Fix docs and add a test showing that we always retry retryable
service acquisition failures `Retries.Effort` times, regardless of
the budget set for RequeueFilter.

RB_ID=914659