finagle-memcached: Fix flakey test in integration.MemcachedTest
Problem

The "re-hash when a bad host is ejected" test in
integration.MemcachedTest is flakey because we can't control
which node (server) the Distributer picks as the backend for
a partitioned client of the hash of a given key (even if we
can control the hashing function); thus, for a given number
of backends, and a client partitioned for N keys,
we can't guarantee 100% usage of the backends, and shutting
one down will not trigger ejection and cache misses if it is not
used by any of the clients.

Solution

Shut down N - 1 of the backends; it's very unlikely that all
partitioned clients hash to the same backend, and thus shutting
down the servers will trigger ejection.

RB_ID=914768